+++ src/PDDLParser/Implementation/Literal.cs
@@ -1,4 +1,5 @@
 using System.Collections.Generic;
+using System.Text;
 
 namespace AIInGames.Planning.PDDL.Implementation
 {
@@ -13,6 +14,33 @@ namespace AIInGames.Planning.PDDL.Implementation
             Predicate = predicate;
             Arguments = arguments;
             IsNegated = isNegated;
+        }
+
+        public string ToPddl()
+        {
+            var sb = new StringBuilder();
+            AppendToPddl(sb);
+            return sb.ToString();
+        }
+
+        internal void AppendToPddl(StringBuilder sb)
+        {
+            if (IsNegated)
+                sb.Append("(not ");
+
+            sb.Append("(");
+            sb.Append(Predicate.Name);
+
+            foreach (var arg in Arguments)
+            {
+                sb.Append(" ");
+                sb.Append(arg);
+            }
+
+            sb.Append(")");
+
+            if (IsNegated)
+                sb.Append(")");
         }
     }
 }